requirements.txt
numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
plotnine
plotly
tqdm
opencv-python
Pillow
pyyaml
ultralytics
sahi
umap-learn
filterpy
pytest